+++ Changelog.md
@@ -1,5 +1,7 @@
 ## 0.0.94
 + Added `muna.beta.openai.images.create` method for using image generation models via our OpenAI-compatible client.
++ Added `beta.Annotations.ImageWidth` annotation method to specify image generation width parameters.
++ Added `beta.Annotations.ImageHeight` annotation method to specify image generation height parameters.
 + Added `beta.Annotations.ImageCount` annotation method to specify image generation count parameters.
 
 ## 0.0.93